Add tests for prom_rate
